requirements.txt
asgiref==3.3.4
attrs==20.3.0
awsebcli==3.19.4
bcrypt==3.2.0
blessed==1.18.0
botocore==1.20.62
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mux-python==1.12.0
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.5.9
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.58.0
whitenoise==5.2.0
